+++ identityhub-service/src/main/java/br/dev/andrestamatto/identityhub/clientapplication/domain/ApplicationClientSettings.java
@@ -1,0 +1,7 @@
+package br.dev.andrestamatto.identityhub.clientapplication.domain;
+
+public sealed interface ApplicationClientSettings
+        permits ProtectedApiSettings, SpaSettings {
+
+    ApplicationClientType type();
+}